[hello-systick-overflow] Add systick overflow example

+++ 08-hello-systick-overflow/src/ticker.rs
@@ -1,7 +1,8 @@
 use cortex_m_rt::exception;
 use stm32f1xx_hal::rcc;
 
-use core::sync::atomic::{AtomicU64, Ordering};
+use portable_atomic::AtomicU64;
+use core::sync::atomic::Ordering;
 use fugit::{Duration, Instant};
 
 pub type TickInstant = Instant<u64, 1, 1_000>; // 1 kHz clock
@@ -12,7 +13,7 @@ static TICKS: AtomicU64 = AtomicU64::new(0);
 use cortex_m::peripheral::syst::SystClkSource;
 
 fn ticks() -> u64 {
-    TICKS.load(Ordering::Relaxed) as u64
+    TICKS.load(Ordering::Relaxed)
 }
 
 pub struct Ticker;
@@ -33,7 +34,7 @@ impl Ticker {
     }
 
     pub fn now(&self) -> TickInstant {
-        TickInstant::from_ticks(ticks() as u64)
+        TickInstant::from_ticks(ticks())
     }
 }
